Add 0.5 offset to OvalRois so pixel origin is always the center
In IJ1 OvalRois have a pixel origin that is the upper left corner of the
pixel, but PointRois have a pixel origin in the center of the pixel. The
pixel origin for integration and fitting is always considered to be the
center. Here we add 0.5 when adding OvalRois to the RoiManager. This is
then later removed if using OvalRois for integration or fitting in other
tools including the MoleculeIntegrator and TransformROIsCommand