Reorganize and update MoleculeIntegrator for custom IntegrationMaps
IntegrationMaps that maps T to UID to Peak can specified together with
the channel, region, and name (usually color name) in headless mode. The
same API is implemented and used during the normal mode. This allows for
drift correction of peak positions that are to be integrated and future
adaptation for quadview setups by allowing for addition of Intervals for
specifying each integration region. Currently region 1 is LONG and
region 0 is SHORT using the provided getter and setter dimension
methods.