Store interval in IntegrationMap and MoleculeIntegrator unit test
Reorganize MoleculeIntegrator to use Intervals provided with each
integrationMap. LONG and SHORT and created from dialog and/or getters
setters, but other could be provided instead or in addition to. This
would allow for processing of quadview microscopy data and provides
clearer workflows.

This commit include completed units with for multicolor FRET and single
color data and confirms correct integration.