Add Location based methods to IOPlugins ensure file opening with common
This ensures that supportsOpen methods using Location exist and are
called. version scijava-common 2.85.0 fails to open files that do not
implement Location methods, so this change ensures that all version of
scijava-common support opening of mars archives and tables.

Locations are not supported in mars, yet, this is just a workaround for
loss of backwards support.