Add C 0 in MoleculeIntegrator if no channels exist to allow processing
The TIRF video provided with "Precision and accuracy of single-molecule
FRET measurements—a multi-laboratory benchmark study" Hellenkamp etl al.
Nature Methods 2018 has not channel information despite the fact that
there are pulses of red and green for donor and acceptor excitation.
This commit fixes a bug with these videos. Not a dummy channel 0 is
always added to ensure no error and that videos are processed.