Update Image names to match Metadata after translation
Some where, probably during translation to OME the image name is
replaced with the file name that was opened. We should reset this so the
correct image name is set. Otherwise, the name could be metadata.txt for
micromanager datasets or the name of the single image that was opened.