Resolves #20 - adds metadata UID source option to image commands
In addition to adding the metadata UID source option to image commands,
this adds a method to MarsOMEUtils for generating unique UIDs. First
attempt is to retrieve the metadata UID from the UUID in the
OMEXMLMetadata record. If that is null, the acquisition date for the
image is retrieved and converted to the UID. If that fails null is
returned and random metadata UIDs are added in the image commands.