Prevent swing EDT issues in PeakFinderCommand so preview overlay draws
Previews are started on their own thread so they can be interrupted if
too much time passes. However, this is not the swing EDT, so it should
not update GUI elements. When attempting sometimes previews did not
display as a result. SwingUtilities.invokeLater is use for all image
updates inside the preview thread to avoid this issue.
- Also, NPE from null JLabels is fixed.