Remove scijava-ui-swing overrides for choiceBoxes, updated in scijava
Removes MarsSwingChoiceWidget that was added to make sure dynamic
updates to dialogs was possible where changing the selection could
change other ChoiceBoxes. But this feature has now been merged into
scijava-ui-swing so the override is not needed anymore.

Removed MarsObjectChoiceWidget that was added to avoid NPE form
swing-ui-scijava. This is now fixed so this resolves #39